[LAUN-12]Rename directories & make story
