+++ src/store/modules/statusBar/lib.ts
