+++ src/store/modules/statusBar/types.ts
