+++ src/views/statusBar/WindowControls.vue
